requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
langchain==0.1.4
langchain-community==0.0.16
langchain-openai==0.0.5  
langchain-groq==0.0.1
pinecone==3.2.2
sentence-transformers==2.3.1
torch==2.1.2
pypdf==4.0.1         
python-docx==1.1.0    
beautifulsoup4==4.12.3
lxml==5.1.0           
pandas==2.2.0         
asyncpg==0.29.0        
sqlalchemy==2.0.25   
alembic==1.13.1        
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1
streamlit==1.31.0
plotly==5.18.0 
httpx==0.26.0
requests==2.31.0
python-json-logger==2.0.7 
tenacity==8.2.3            
tiktoken==0.5.2            
pytest==8.0.0        
pytest-asyncio==0.23.4 
black==24.1.1          
ruff==0.2.0           